% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/didint-package.R
\docType{package}
\name{didint-package}
\alias{didint}
\alias{didint-package}
\title{didint: Difference-in-Differences with Spatial Interference}
\description{
Doubly robust estimators for difference-in-differences when potential outcomes depend on neighbours' treatment status (spatial interference). Implements the 2x2 case of Xu (2023) and the dynamic / staggered extensions of Xu (2026), with spatial-HAC standard errors based on unit locations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://xiangao.github.io/didint/}
  \item \url{https://github.com/xiangao/didint}
  \item Report bugs at \url{https://github.com/xiangao/didint/issues}
}

}
\author{
\strong{Maintainer}: Xiang Ao \email{xiangao@gmail.com}

Authors:
\itemize{
  \item Xiang Ao \email{xiangao@gmail.com}
}

}
\keyword{internal}
